Workaround: Terminal windows are highlighted, even though 'buftype' is set
But too late; the terminal buffer is first created as a normal empty buffer ([No Name], with empty 'buftype'). Only later :setlocal buftype=terminal happens (and the OptionSet event also isn't triggered for it, so we cannot use that to hook into). Fortunately, a separate TerminalOpen event was created.
As the default filter is defined in a separate autoload script, we can simply define the autocmd during sourcing of the script.