Compatibility: Ensure that highlight is visible
Some colorschemes use red in foreground to show Error, and leave
background the same as Normal. In those cases, the current
implementation wouldn't work.

This commit checks for the foreground and background colors of the Error
highlight group individually and sets the appropriate background color
for our ShowTrailingWhitespace highlight group.